refactor: Introduce DataFrames
This should reduce the amount of work during logging. Only the required information is added to `DataFrame`s,
which are then rendered into json by the underlying json engine.

They might introduce some memory pressure, due to the creation of more instances (notably, the message DataFrame), but that
need to be measured and timed. If it does indeed add pressure to the memory, it can be solved by object pools.