refactor: Make logger creation lazy
This is likely a huge improvement over the previous implementation. This one change allows us to:
  - Remove a few complicated functions in favor of simpler alternatives;
  - Ensure clients of the TreeCache don't need to create Loggers beforehand;
  - Lazily create the loggers, only when requested;
  - Define a much more cohesive API for the clients.

In a nutshell, the new API is simpler. You operate on Configs and identifiers, mostly.
You only get loggers when you request them.
This makes the TreeCache slightly more specialized, as it has to know how to create MapleLoggers,
but that is not a problem because it is already too bound to MapleLoggers anyway.

Also, we ended up removing all the generic Loggers we were returning, becasue inside Maple we don't
need to guess - it should always be MapleLoggers.

Then, the change allows us to simplify the search and base everything in the getOrCreate method
(previously createRecursively), which is idempotent now (due to fixing a bug), making it perfect
for internal usage over the previous search method. This ensures all updateConfig methods will
effectively write to the cache and no extra loggers will every be created.

They're only created on-demand by the MapleLoggerFactory.