refactor: Simplify setting up and updating config
In essence, there's always a config value, so all changes to the config will be
some sort of reconfiguration. Hence, there's no real value in shipping on api for
'setting' and one for 'updating' the config. If one wants to ignore the existing
value, one can just ignore the argument in the function.

So this not only makes the API simpler but also reduces a lot of the duplication,
which might make it confusing for users.