refactor: Move the precision to an external variable
In the future, it can even be that we might want to tweak this value.
If we make it smaller, it'll be more precise at the expense of some CPU processing;
If we make it bigger, it'll be less precise, but spare CPU;

It is important to note that 1ms interval is not necessarily correct in terms of precision given, in linux,
System.currentTimeMillis() takes several hundred useconds.