refactor: Simplify thread init logic
Using a lock instead of synchronized allows us to simplify the logic by skipping the second check.
While this is only useful during thread initialization, it is vital to avoid race conditions.

The alternative would be to always start the thread, which might be excessive if we're not using
timestamp