refactor: Use ByteBuffers to write to stdout
This is a necessary fix for JDK19 (tested) because of how FileWriter writes to the stdout in the newer
jdks. Apparently, either the buffer size changed or an internal write ceased to be synchronized.
Either way, the outcome is that the output is now garbled when on multi-threaded environment.

To mititage the issue (for now, until 0.6 at least) we introduce the ByteBufferWriter, which
will act as a bridge between the FileChannel created by `FileOutputStream` and the `ByteBuffer`.

It is very simple and might require some tunning, but nonetheless works fine to wrap the byte[] with a
CharBuffer and write that to the ByteBuffer before flushing the contents to stdout through the channel atomically.

On a future version, we definitely want to revisit this, but it should be enough to ensure integrity of the
log messages, effectively solving the issue reported by #28.