feat: Add throwable format
Also, hidden in this commit, is a bloom filter implementation to ensure we're not
writing the same stacktrace multiple times. A bloom filter should be OK here because,
even though we don't want false-positives, the filter is relatively small (~1kb), which is
more than enough to have a minimum collision rate.

In fact, for m=8192 and k=6, we have 2% collision rate if we reach maximum stack size (1024, usually).
At a fairly more common, though still very big stack of size 256, we have p=0.0000249 (1 in 40.000)
and at a stack size of mere 128 frames we're almost at 1 in every 2 million.

So, in other words, the bloom filter implementation using the native hashCode with a Swamidass & Baldi
algorithm for k-hashes, will be space-efficient (~1kb) and fast enough for us to have a decent cache for our
stack-trace dedupe.