refactor: Break dependency on the Config object
This allows for a small runtime optimization, where the sink doesn't
have to a) check whether filter is enabled or not and b) update state.
Because this object is local (and unique) per logger, it is only updated
when the config changes, re-calculating the values if needed.

This opens up the possibility for having more flexible sink for better
different message layouts, for example.