refactor: Add OutputManager to manage out channels
This is not strictly needed, but it is a nice refactoring towards a
better desing, in which we can setup and configure output channels for
test differently than the ones we have "live"

This is still "wip", but it is in a good enough shape to be deployed.

This refactoring also allows us to further down the road simplify
`DirectJson` and `SinkImpl` by handling all the channel-related stuff
there.