refactor: make key strings faster
In general, key strings should not contain escape sequences or unicode
chars, so this is a good candidate for optimization, as we can skip
checking individual characters for those traits. Furthermore, we can
certainly append it with the `:` character, not needing to overwrite
previously added `,` character by reusing the `writeString` method.