refactor: Avoid copying a context map when logging
It is expensive in terms of memory to allocate a new map just to hold
the contents of data in the map.

Given the MDC adapter stores data in a tree-like data structure, we
would need two operations:
1) traverse back to the root node and populate the context map,
extending it at every node down until the leaf;
2) create an EntrySet to iterate over the items;

This implementation drops the need for an intermediary map and abstracts
away the tree structure.