fix: PLaceholder escaping logic
There are many layers in this commit:

1. The new logic is arguably "clearer" and "easier to understand";
2. The new logic is more correct; and
3. The new logic is more performant.

1. Cleaner:

We already know when we put an escape in the string.
A double-escape invalidates a single-escape in terms of replacing the
placeholder, so moving the check to when we put the escape makes more
sense then.

2. More correct:
There are a few bugs in the previous logic:
- Arguments can be over and we still have placeholders;
- `buffer.put(buffer.position() - offset, DELIM_START);` causes chars to
  go missing;
- `null` arguments should be ignored and the placeholders untouched;
- escaped placeholders should not leave `\` behind.

3. More performant:
Virtually no difference in the benchmark, but logically it makes some
sense:
- We have more placeholders than usually we have escapes, so we don't
  need to constantly check the string for escapes.
- We do fewer checks (not necessarily a reason for being faster, but
  still);