refactor: Replace ThreadLocal with object pool
This also flips the logic a little. Before, we would ensure each thread
had a pool of `PennaLogEvent`, and those would then be forwarded to the
Sink bound to that thread-local instance.

Now a `PennaLogEvent` and a `Sink` sit side-by-side in the event pool in
the `LogUnitContext`. That is used by the guards to write the first to
the latter and immediately returns itself to the object pool.

The `ObjectPool` is currently hardcoded to have only 16 elements. Future
improvements can be made to arrange larger values if needed.