feat: Add new logger cache implementation
The tree-based implementation is very fast to update but not much to
create new loggers. Furthermore, it might not be space efficient, though
I haven't checked that thoroughly.

One final word about it is that it could also be that fast because it is
not necessarily thread safe.

This implementation is more adequate because it is faster for creating a
new logger than the tree-based version.

It is likely more space efficient than the previous one as it doesn't
have much acessory code around it.