refactor: Use new LoggerStorage in LoggerFactory
This comes with a surprising and very positive change. Now, it is not
necessary to break the logger by `.` using a regex as the internal code
of the storage handles that (without a regex), so we're left with better
APIs and a smaller footprint for the LoggerFactory. Also, given we
lifted the requirement of `String[]` in the config update code, we don't
need to do any special-casing for the records, so they're now able to
use the native methods, which makes the code much cleaner to maintain
now.