refactor: Rework internal TST navigation
There are quite a few changes packed here, so I'll try to summarize:

- Refactor the `ComponentNode` records into a simple `Node` class:

This is important because we unlock a simpler mutability (no need to
have atomic references), simpler creation (the only argument required is
the component) and we remove the redundant control over updates (atomic
references vs locks, in favour of locks).

- Replace char[][] structure with a simple String

Altough the original idea was to save memory by using char[][], it was
actually much worse as proven by a few tests, so instead we work with
strings natively.

- Drop cursors in favour of StringNavigator

The StringNavigator is an improvement over the Cursors as we can keep
the search code local to the functions that require it. For the
comparison of the components in a string we rely in the
`StringNavigator` and the `StringView` classes, which create much less
garbage. The `StringView` object just clips the original string by
adding an offset based on `.` and `$` positions in the base string.