refactor: Use TreeMap for the MDC
TreeMap seems to be relatively lighter than HashMap (very minimally) and
performance is similar.

The reason for dropping the old implementation is to reduce complexity,
so instead of a complex in-house implementation we opt for a simpler
layer that outsources to the TreeMap.

Future iterations can be made by playing around with the `StringMap`
interface.