feat: Add support for Snake Yaml {,Engine}
With this commit the yaml config module now requires to have one of the
three dependencies so it can properly read the yaml configuration.

More libraries should come but those three are arguably the most popular
ones for now so it makes sense to allow for something that the users
might be already using.