refactor: Cleanup old configuration + penna-dev
The old configuration mechanism is effectively replaced by the new
one. This allows us to remove a lot of unnecessary methods, classes and
interfaces and the penna-dev project.

The latter had two different justifications for being developed in the
first place:
 - Allow for a development-runtime that shows up the logs in
   human-readable format. That can be much more effectively be replaced
   by fblog or other alternatives, so it doesn't stand on its own here;
 - Allow for runtime level tweaking, so tests can change the verbosity
   of the logs if necessary. This is effectively replaced by the new
   config interfaces, so that in itself also doesn't sustain a strong
   reason for maintaining this module.

Therefore, it is better to remove it not to bloat penna, so the surface
can remain small and the focus can be on the important features.